requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
trafilatura==1.9.0
tavily-python==0.3.3
openai==1.40.0
python-dotenv==1.0.1
